v3.4.2

+++ UI/Models/ProduktModel.cs
@@ -7,7 +7,7 @@
         public string Title { get; set; }
         public int? QuantityInStock { get; set; }
         public string Location { get; set; }
-        public double? Price { get; set; }
+        public float? Price { get; set; }
         public string Description { get; set; }
     }
 }
